Improved highlight extraction & explicit defaults
Calls to extract only a single highlight group has been significantly
improved. Instead of relying on possibly outdated information, it
directly gets the information from the `:highlight` command itself. This
may be a bit slower than searching a dictionary, but the less frequent
calls to generate the dictionary in the first place should be a nice
boost in speed.

Furthermore, refreshing the undefined highlight groups is now only done
when in debug mode.

Finally, added a set of highlight groups (for now `Number` and `Float`)
that I have explicitly stated can take on their default values set by
the editor.